komodo_notaryvin: get rid of ugly parameter passing via void pointer
to pass two params in a function earlier we used a (void **) to
handle pointers on these 2 params in "array". also there was "hidden"
parameter, this pointer itself, if pointer == 0 -> it before HF time,
if pointer !=0 -> after HF time, and we should extract two params from
array of passed pointers. of course it can be replaced with tuple, or
boost::optional with tuple "inside", but much more simple way is to
pass params (CScript and nLockTime) in komodo_notaryvin same way as they
exits and assume that if passed CScript().size() == 0, it before HF
and if size() > 0 - it after HF.